Fix intraday persistence crash on partial OHLC bars
Real-world HOURLY SRAZ26 download (539 bars) failed at
database.cache.insert_bars() with:
  TypeError: float() argument must be a string or a real number, not 'NAType'

Root cause: LSEG can legitimately return a thin intraday bar (e.g. no
trade printed within that hour) with one or more OHLC fields missing
while others (commonly Close) are populated. That missing value can
surface as pd.NA (a pandas nullable-dtype sentinel) rather than
np.nan, and insert_bars called Python's float() directly on each raw
cell -- float() accepts np.nan but raises on pd.NA.

Fix, per the approved architecture (preserve genuine missingness as
SQL NULL; never fabricate or drop):

- database/models.py: PriceBar.open/high/low/close/volume are now
  individually nullable. ric/interval/datetime remain mandatory.
- core/downloader.py: _normalize_columns explicitly casts each OHLCV
  column to plain numpy float64 after pd.to_numeric, so pd.NA never
  survives past normalization -- missing values become np.nan, never
  filled, never dropped.
- database/cache.py: insert_bars now converts each cell via
  `None if pd.isna(value) else float(value)`, so pd.NA/np.nan/None
  all persist as SQL NULL. Verified (and left unmodified, since
  testing confirmed it already works) that read_bars reconstructs
  SQL NULL back into np.nan in a canonical float64 column.

SyncRange semantics, Strategy Engine, and Module 4A are unchanged --
a successfully downloaded period may legitimately contain bars with
missing fields, and a bar with a missing Open but valid Close remains
usable for a Close-based strategy exactly as before.

No migration framework added. CHANGELOG documents that a pre-existing
local data/oscill8.db (built under the old NOT NULL schema) must be
deleted and rebuilt once -- it's a pure, fully re-fetchable LSEG
cache, so nothing is permanently lost. No database file was touched
by this change.

18 new/changed tests across test_downloader.py, test_cache.py,
test_models.py, test_service.py.

+++ core/downloader.py
@@ -149,7 +149,15 @@ def _normalize_columns(df: pd.DataFrame) -> pd.DataFrame:
     out = out[_CANONICAL_COLUMNS]
     out["Date"] = pd.to_datetime(out["Date"])
     for col in ["Open", "High", "Low", "Close", "Volume"]:
-        out[col] = pd.to_numeric(out[col], errors="coerce")
+        # pd.to_numeric preserves a pandas nullable extension dtype (e.g.
+        # "Float64") if the input already has one, which represents a
+        # missing value as pd.NA rather than np.nan. LSEG can legitimately
+        # return such a column for a thin intraday bar (e.g. an hourly
+        # bucket with no trade printed) -- explicitly cast to plain numpy
+        # float64 so every downstream consumer only ever sees np.nan for
+        # a missing value, never pd.NA. The missing value itself is
+        # preserved, not filled or dropped.
+        out[col] = pd.to_numeric(out[col], errors="coerce").astype("float64")
 
     return out.sort_values("Date").reset_index(drop=True)
